Integrate plots and render into single panel
